backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pytesseract
pillow
pymupdf
transformers
torch
requests
python-multipart
